+++ config/routes.rb
@@ -1,26 +1,28 @@
 # frozen_string_literal: true
 
 Rails.application.routes.draw do
-  get 'external/home'
+  get "external/home"
 
-  root 'welcome#index'
+  root "welcome#index"
 
-  get 'pc' => 'pc#index'
+  get "pc" => "pc#index"
+
+  resources :directors
 
   namespace :pc do
     get :clients
     get :contact
     get :index
-    get :our_work, path: 'our-work'
-    get :director_christian_williams, path: 'directors/christian-williams'
-    get :director_kamell_allaway, path: 'directors/kamell-allaway'
-    get :director_noah_lydiard, path: 'directors/noah-lydiard'
-    get :director_vladimir_minuty, path: 'directors/vladimir-minuty'
+    get :our_work, path: "our-work"
+    get :director_christian_williams, path: "directors/christian-williams"
+    get :director_kamell_allaway, path: "directors/kamell-allaway"
+    get :director_noah_lydiard, path: "directors/noah-lydiard"
+    get :director_vladimir_minuty, path: "directors/vladimir-minuty"
   end
 
   resources :videos do
     collection do
-      get 'reorder/:video_page', action: 'reorder', as: :reorder
+      get "reorder/:video_page", action: "reorder", as: :reorder
       post :save_video_order
     end
   end
@@ -76,67 +78,69 @@ Rails.application.routes.draw do
     end
   end
 
-  get 'invite/:invite_token' => 'project_users#invite'
+  get "invite/:invite_token" => "project_users#invite"
 
-  get '2016', to: redirect('year-in-review')
-  scope module: 'welcome' do
+  get "2016", to: redirect("year-in-review")
+  scope module: "welcome" do
     get :dashboard
     get :index
     get :work
     get :drtv
-    get 'images/videos/:video_id', action: :download_video_image, as: :download_video_image
+    get "images/videos/:video_id", action: :download_video_image, as: :download_video_image
     get :contact
     post :submit_contact
-    get :year_in_review, path: 'year-in-review'
+    get :year_in_review, path: "year-in-review"
     # get :news
     # get :about
     # get :clients
     # get :index_v1
   end
 
-  scope module: 'application' do
+  scope module: "application" do
     get :version
   end
 
   scope module: :external do
     get :landing_draft, path: "landing/draft"
+    get :director, path: "director/:director", as: :public_director
+    get :photo_director, path: "director/:director/photo"
   end
 
-  scope module: 'client_session' do
-    get  'client/login',  action: 'new',     as: :client_login
-    post 'client/login',  action: 'create'
-    get  'client/logout', action: 'destroy', as: :client_logout
+  scope module: "client_session" do
+    get  "client/login",  action: "new",     as: :client_login
+    post "client/login",  action: "create"
+    get  "client/logout", action: "destroy", as: :client_logout
   end
 
   devise_for :users,
-             controllers: { sessions: 'sessions' },
+             controllers: { sessions: "sessions" },
              path_names: {
-               sign_up: 'join',
-               sign_in: 'login',
-               sign_out: 'logout'
+               sign_up: "join",
+               sign_in: "login",
+               sign_out: "logout"
              },
-             path: ''
+             path: ""
 
   resources :users
 
   if Rails.env.development?
-    get '/rails/mailers' => 'rails/mailers#index'
-    get '/rails/mailers/*path' => 'rails/mailers#preview'
+    get "/rails/mailers" => "rails/mailers#index"
+    get "/rails/mailers/*path" => "rails/mailers#preview"
   end
 
-  scope module: 'client_project' do
-    get ':id', action: 'root', as: :client_project_root
-    get ':id/menu', action: 'menu', as: :client_project_menu
-    get ':id/agency_logo', action: 'agency_logo', as: :client_project_agency_logo
-    get ':id/client_logo', action: 'client_logo', as: :client_project_client_logo
+  scope module: "client_project" do
+    get ":id", action: "root", as: :client_project_root
+    get ":id/menu", action: "menu", as: :client_project_menu
+    get ":id/agency_logo", action: "agency_logo", as: :client_project_agency_logo
+    get ":id/client_logo", action: "client_logo", as: :client_project_client_logo
 
-    get ':id/documents/download/:document_id', action: 'download_document', as: :client_project_download_document
-    get ':id/documents/download_primary/:document_id', action: 'download_primary_document', as: :client_project_download_primary_document
-    get ':id/photos/download/:gallery_photo_id', action: 'download_gallery_photo', as: :client_project_download_gallery_photo
-    get ':id/:top_level/:category_id/photos/:gallery_id', action: :gallery_show, as: :client_project_gallery
-    get ':id/:top_level/:category_id/photos/:gallery_id/:gallery_photo_id', action: :gallery_photo, as: :client_project_gallery_photo
-    get ':id/:top_level/:category_id', action: 'category', as: :client_project_category
-    get ':id/:top_level/:category_id/:document_id', action: 'document', as: :client_project_category_document
-    get ':id/:category_id', action: 'category'
+    get ":id/documents/download/:document_id", action: "download_document", as: :client_project_download_document
+    get ":id/documents/download_primary/:document_id", action: "download_primary_document", as: :client_project_download_primary_document
+    get ":id/photos/download/:gallery_photo_id", action: "download_gallery_photo", as: :client_project_download_gallery_photo
+    get ":id/:top_level/:category_id/photos/:gallery_id", action: :gallery_show, as: :client_project_gallery
+    get ":id/:top_level/:category_id/photos/:gallery_id/:gallery_photo_id", action: :gallery_photo, as: :client_project_gallery_photo
+    get ":id/:top_level/:category_id", action: "category", as: :client_project_category
+    get ":id/:top_level/:category_id/:document_id", action: "document", as: :client_project_category_document
+    get ":id/:category_id", action: "category"
   end
 end