+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - **General Changes**
   - Login cookies are now cross subdomain and work between www and non-www URLs
   - The DRTV page now redirects to the Work page
+- **Director Changes**
+  - Admins can now create and edit director pages
 - **Video Changes**
   - Admins can now set Client Name and Agency Name for videos
   - Hovering over videos now displays Client Name and Agency Name